scale-generator: WIP extract API from example implementation
