+++ exercises/scale-generator/Cargo-example.toml
@@ -1,0 +1,10 @@
+[package]
+name = "scale_generator"
+version = "1.0.0"
+
+[dependencies]
+enum-primitive-derive = "^0.1"
+failure = "0.1"
+failure_derive = "0.1"
+itertools = "0.7"
+num-traits = "^0.1"
