+++ exercises/scale-generator/Cargo.toml
@@ -3,8 +3,4 @@ name = "scale_generator"
 version = "1.0.0"
 
 [dependencies]
-enum-primitive-derive = "^0.1"
 failure = "0.1"
-failure_derive = "0.1"
-itertools = "0.7"
-num-traits = "^0.1"
